+++ mayan/apps/forms/forms.py
@@ -17,9 +17,7 @@ from .form_fields import (
     CharField, ChoiceField, ModelChoiceField, ModelMultipleChoiceField,
     MultipleChoiceField
 )
-from .form_mixins import (
-    FormMixinDynamicFields, FormMixinFieldsets, FormMixinFormMeta
-)
+from .form_mixins import FormMixinDynamicFields, FormMixinFieldsets
 from .form_options import DetailFormOption, FilteredSelectionFormOptions
 from .form_widgets import (
     DisableableSelectWidget, PlainWidget, RadioSelect, Select, SelectMultiple,
@@ -27,11 +25,11 @@ from .form_widgets import (
 )
 
 
-class Form(FormMixinFormMeta, FormMixinFieldsets, DjangoForm):
+class Form(FormMixinFieldsets, DjangoForm):
     """Mayan's default form class."""
 
 
-class ModelForm(FormMixinFormMeta, FormMixinFieldsets, DjangoModelForm):
+class ModelForm(FormMixinFieldsets, DjangoModelForm):
     """Mayan's default model form class."""